UniqueItem is not triggered when completion is initiated by backspace
